Create unit tests for LedgerWriterController
Create unit tests for methods versions() and readiness() in the
LedgerWriterController class.

Refactor LedgerWriterController constructor so that JWTVerifier and
environment valuables can be injected during testing. Wrap JWTVerifier
into a class and create a custom runtime exception for the class when
it fails to generate the verifier.

Create a public final variable READINESS_CODE for "ok" so that the
unit tests can refer to the variable instead of the string.

Add Google license header

Fix Java style

Fix Java style